seperate parquet reader and arrow convert
